backend/requirements.txt
flask
flask-pymongo
pyjwt
bcrypt
python-dotenv
flask-cors
flasgger
pytest
requests
